Refs 144, splits maintenance/forms.py in severals files
